Copying urls on history page also shows a toast notification
